Centralize plugin validation
