+++ src/core/plugins.py
@@ -1,12 +1,10 @@
 import importlib.util
 from pathlib import Path
-from typing import Optional
 
 import requests
 from loguru import logger
-from pydantic import BaseModel, Field, ValidationError, field_validator
 
-from core.interfaces import IParser
+from core.interfaces import IParser, validate_parser, class_variables
 from core.settings import settings
 
 
@@ -44,36 +42,37 @@ def download_plugin(file_type: str, plugin_name: str, save_path: Path):
         print(f"Error downloading plugin {plugin_name}: {e}")
 
 
-class PluginMetadata(BaseModel):
-    plugin_name: str
-    statement_type: str
-    version: str
-    description: Optional[str] = Field(
-        "", description="A short description of the plugin."
-    )
-    search_string: str
-    file_path: str
+def load_plugin(plugin_file: Path) -> tuple[str, IParser, dict[str, str]]:
+    """
+    Dynamically load the Parser class from a plugin module, validate it, and retrieve metadata.
+    plugin_file.name like 'pdf_citicc_v0.1.0.pyc'
+    """
+    # Load the module from file
+    plugin_name = plugin_file.stem
+    spec = importlib.util.spec_from_file_location(plugin_name, plugin_file)
+    if not spec or not spec.loader:
+        raise ImportError(f"Cannot load module from {plugin_file}")
+    module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(module)
 
-    # Validator to reject empty string values
-    @field_validator(
-        "plugin_name", "statement_type", "version", "search_string", mode="before"
-    )
-    def no_error_or_empty(cls, value, field):
-        """
-        Ensure no field is an empty string.
-        """
-        # field_name = field.field_name
-        if value == "":
-            raise ValueError(f"Field '{field.field_name}' cannot be empty.")
-        return value
+    # Ensure this module contains a Parser(IParser)
+    ParserClass = getattr(module, "Parser", None)
+    if not ParserClass:
+        raise ValueError(f"No 'Parser' class found in {plugin_file}")
+    if not isinstance(ParserClass, IParser):
+        raise TypeError(f"Plugin {plugin_name} must implement IParser")
 
-    # Validator to set default description
-    @field_validator("description", mode="before")
-    def default_description(cls, value):
-        """
-        If description is empty, set a default value.
-        """
-        return value or "No description provided."
+    # Validate that the plugin overrides required class variables (metadata)
+    required_variables = class_variables(IParser)
+    try:
+        validate_parser(ParserClass, required_variables)
+    except Exception as e:
+        raise ImportError(f"Failed to load plugin {plugin_name}: {e}")
+
+    # Extract plugin metadata
+    metadata = {var: getattr(ParserClass, var) for var in required_variables}
+
+    return plugin_name, ParserClass, metadata
 
 
 class PluginManager:
@@ -88,59 +87,15 @@ class PluginManager:
         self.plugins = {}
         self.metadata = {}
 
-        for plugin_file in settings.plugin_dir.glob("**/*.pyc"):
-            # plugin_name like 'pdf.citibank'
-            module_name = plugin_file.stem
-            if module_name.split(".")[0] == "__init__":
-                continue
-            file_type = plugin_file.parent.name
-            plugin_name = ".".join([file_type, module_name])
-
+        for plugin_file in settings.plugin_dir.glob("*.pyc"):
             # Retrieve the Parser(Iparser) class from the plugin and store it
-            ParserClass, metadata = self.load_module_from_file(plugin_name, plugin_file)
-            if ParserClass:
+            try:
+                plugin_name, ParserClass, metadata = load_plugin(plugin_file)
                 self.plugins[plugin_name] = ParserClass
                 self.metadata[plugin_name] = metadata
-
-    def load_module_from_file(self, plugin_name: str, plugin_file: Path):
-        """
-        Load a Python module from a .pyc file.
-        """
-        try:
-            # Load the module and spec
-            spec = importlib.util.spec_from_file_location(plugin_name, plugin_file)
-            if not spec or not spec.loader:
-                raise ImportError(f"Cannot load module from {plugin_file}")
-            module = importlib.util.module_from_spec(spec)
-            spec.loader.exec_module(module)
-
-            # Get the class called Parser(IParser) from the module
-            ParserClass = getattr(module, "Parser", None)
-            if not ParserClass:
-                raise ValueError(f"Parser class not found in {plugin_file}")
-            if not isinstance(ParserClass, IParser):
-                raise TypeError(f"{ParserClass} must implement IParser")
-
-            # Extract and validate the metadata
-            metadata = {
-                "plugin_name": plugin_name,
-                "statement_type": getattr(ParserClass, "STATEMENT_TYPE", ""),
-                "version": getattr(ParserClass, "VERSION", ""),
-                "description": getattr(ParserClass, "DESCRIPTION", ""),
-                "search_string": getattr(ParserClass, "SEARCH_STRING", ""),
-                "file_path": str(plugin_file),
-            }
-
-            # Validate metadata
-            metadata = PluginMetadata(**metadata)
-
-            return ParserClass, metadata
-        except ValidationError as e:
-            logger.error(f"Validation error for plugin {plugin_name}: {e}")
-            return None, None
-        except Exception as e:
-            logger.error(f"Failed to load plugin {plugin_name}: {e}")
-            return None, None
+                logger.success(f"Loaded plugin: {plugin_file}")
+            except Exception as e:
+                logger.error(f"Failed to load {plugin_file}: {e}")
 
     def get_parser(self, plugin_name: str):
         """
